feat(garden): Split SunflowersHud from AccountHud
Closes #119
