+++ packages/game/src/hooks/useCurrentGarden.ts
@@ -1,0 +1,6 @@
+export function useCurrentGarden() {
+    return {
+        data: { id: '1', name: 'Moj vrt' },
+        isLoading: false
+    };
+}
