+++ packages/game/src/hooks/useSunflowers.ts
@@ -1,0 +1,8 @@
+export function useSunflowers() {
+    return {
+        data: {
+            activity: []
+        },
+        isLoading: false
+    }
+}
