+++ packages/game/src/shared-ui/NoNotificationsPlaceholder.tsx
@@ -1,0 +1,10 @@
+import { Typography } from "@signalco/ui-primitives/Typography";
+
+export function NoNotificationsPlaceholder() {
+    return (
+        <Typography level="body2" className="flex items-center gap-2">
+            <span className="text-2xl pb-3 leading-none">📪</span>
+            <span>Nema novih obavijesti</span>
+        </Typography>
+    );
+}
